QPdfFile: allow Qt Quick to change thread affinity to its own renderer
The thread that calls QQuickPixmapReader::processJob() pulls out
the QPdfFile pointer, passes it to readImage() and expects that it
will not become a dangling pointer while readImage() is busy.
So it needs to call moveToThread() to "pull" the affinity to itself;
and that's only possible if it is set to null in the same thread
that initially has affinity.

Task-number: QTBUG-128875
Change-Id: Id54803783f353631f97c7676c90d5b0fbe09f4eb
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>