Add QWebEngineProfileBuilder class
This is a wrapper designed to handle profile construction.

The QWebEngineProfile class allows the modification of profile storage
properties, which is against chromium profile design. Changing storage
properties requires the recreation of storage and network context
leading to hacks, which in turn leads to bugs.

The QWebEngineProfileBuilder aims to prevent modification of storage
properties and set them at the profile creation time. The current change
introduces new APIs without deprecating the storage setters in the
QWebEngineProfile class. The final goal is to deprecate the storage
setters and constructor from QWebEngineProfile and use the
QWebEngineProfileBuilder to create profiles.

Task-number: QTBUG-66068
Change-Id: I3c2562c9e1445708f0dbdeae5783a8c7d38313af
Reviewed-by: Michael Brüning <michael.bruning@qt.io>