Fix for broken rendering when reparenting view
When the dirty scene nodes are updated after reparenting the view, the
texture is not ready and as result rendering is broken. The parent
change triggers QQuickItem::ItemSceneChange, which in turn calls the
RenderWidgetHostViewQtDelegateClient::visualPropertiesChanged(). Even
though the window geometry is changed, this wouldn’t lead to generating
new texture, as the call to synchronize the visual properties is
missing. Invoked synchronizeVisualProperties to avoid such issues.

Fixes: QTBUG-131304
Pick-to: 6.8
Change-Id: I29e6934c33b3fc9f54bfc6e17b29e5ea736cfee0
Reviewed-by: Peter Varga <pvarga@inf.u-szeged.hu>