Add coin sanity check
Our repository hosts two projects and therefore
configure can not fail in case only one of the modules
should be built. This however ends badly if
there is change, which accidentally changes
some dependency checks, in the way that on
some platform it is later skipped without the committer
being aware of the issue.

Therefore add white list, which hardcodes all the nodes
and expected results. This way changes with side
effects of making some coin node not longer supported,
have to be reflected in "coin.nodes" file.

Pick-to: 6.9
Change-Id: I583b42f979db5bf8330d0c0a690f0a27ac712e76
Reviewed-by: Michael Brüning <michael.bruning@qt.io>