Fix spurious Wayland+OpenGL assert
The assert at the end of texture() is intended to catch programming
errors which are our responsibility, but also aborts on a minor invalid
operation in qtbase. Log and clear openGL error flags at the start of
the function to limit the assert to guarding only our code.

This also introduces getGLErrorString to log the flag names.

Fixes: QTBUG-132608
Pick-to: 6.9
Change-Id: I6e7cb9e61757d0290c9cf1bfa5b98a4ffb3e1d52
Reviewed-by: Peter Varga <pvarga@inf.u-szeged.hu>