Use ldd linker in case of clang-cl
By default qtbase does not select any linker,
it only supports explicit selection by linker INPUT
feature to any of bfk , gold or lld linkers.
In case user does not select any linker is it
up to gcc driver or select default one.
Moreover, linker features are explicitly disabled on windows.
However, for webengine it is not a optional feature,
as it switches between different linker calls
on chromium sides.

Enable ldd-linker in case of clang-cl, as this what
we should use for linking for tools on chromium side.

In practice it calls lld-link.exe instead of link.exe on
3rdparty build.

Pick-to: 6.9
Change-Id: I30b55c0efb8a09abdae759edfc301fbccd496f8a
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>