Fix flaky tst_QWebEngineDownloadRequest auto tests
Some of the test cases rely on context menu triggered by simulated user
action. Input events are suppressed by blink if the page is not painted
yet because the cc compositor deferred a commit.
See SuppressingInputEventsBits::kDeferCommits and
WidgetInputHandlerManager::DispatchEvent().

Wait for the HTML <body> element to be painted to make sure the
simulated mouse events will trigger the context menu.

Pick-to: 6.8 6.9
Change-Id: Ic41c03fbceb4a12ff90fc601560d48a50db6488c
Reviewed-by: Moss Heim <moss.heim@qt.io>
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>