Fix flaky tst_QWebEngineView::keyboardFocusAfterPopup on macOS
On macOS, Qt closes all popups when NSWindowDidResignKeyNotification
event is triggered by the OS, see QTBUG-105474.

This event somehow is triggered after we call show() on a widget. The
event might be triggered while the QCompleter's popup is already open
and it breaks the test.

Wait for the OS events to be processed with qWaitForWindowExposed().

Pick-to: 6.8 6.9
Fixes: QTBUG-133590
Change-Id: I51b05df375046fcfb28ae213173a517d60216ae2
Reviewed-by: Anu Aliyas <anu.aliyas@qt.io>