Perform the focus check earlier
While loading URLs, QtWebEngine will remove the old widget from the
layout and then add the new widget. When hiding the old widget from the
layout, it loses focus. As a result, the focus is not set to the new
widget. Modified the code to check whether the old widget has focus
before it is hidden.

Amends e227bbddbfc03c45735978f5b83994235225569a

Fixes: QTBUG-133649
Pick-to: 6.9 6.8
Change-Id: Iac4b1317ed469bfa32e18f0e18361a2279b77cf6
Reviewed-by: Michael Brüning <michael.bruning@qt.io>